refactor: classes to dto + directory structure
