+++ polls/business/dtos/poll_option_dto.py
@@ -1,4 +1,4 @@
-class PollOption:
+class PollOptionDto:
     """
     Poll Option. One of the choices user may 
     choose from.
