+++ polls/business/exceptions/poll_option_unvalid_exception.py
@@ -1,0 +1,8 @@
+class PollOptionUnvalidException(Exception):
+    """
+    A PollOption is unvalid for a certain Poll
+    """
+
+    def __init__(self, *args: object) -> None:
+        super().__init__(*args)
+
